+++ src/Controllers/FilePreviewHandler.php
@@ -2,49 +2,16 @@
 
 namespace Livewire\Controllers;
 
-use Illuminate\Support\Facades\Validator;
 use Livewire\FileUploadConfiguration;
 
 class FilePreviewHandler
 {
+    use CanPretendToBeAFile;
+
     public function handle($filename)
     {
         abort_unless(request()->hasValidSignature(), 401);
 
-        dd($filename);
         return $this->pretendResponseIsFile(FileUploadConfiguration::storage()->path(FileUploadConfiguration::directory().$filename));
-    }
-
-    public function pretendResponseIsFile($file)
-    {
-        $expires = strtotime('+1 year');
-        $lastModified = filemtime($file);
-        $cacheControl = 'public, max-age=31536000';
-
-        if ($this->matchesCache($lastModified)) {
-            return response()->noContent(304, [
-                'Expires' => $this->httpDate($expires),
-                'Cache-Control' => $cacheControl,
-            ]);
-        }
-
-        return response()->file($file, [
-            'Content-Type' => 'application/javascript; charset=utf-8',
-            'Expires' => $this->httpDate($expires),
-            'Cache-Control' => $cacheControl,
-            'Last-Modified' => $this->httpDate($lastModified),
-        ]);
-    }
-
-    protected function matchesCache($lastModified)
-    {
-        $ifModifiedSince = $_SERVER['HTTP_IF_MODIFIED_SINCE'] ?? '';
-
-        return @strtotime($ifModifiedSince) === $lastModified;
-    }
-
-    protected function httpDate($timestamp)
-    {
-        return sprintf('%s GMT', gmdate('D, d M Y H:i:s', $timestamp));
     }
 }
